discord-bot/pyproject.toml
# (C) 1998-2026 Miroslav Sotek. All rights reserved.

[project]
name = "director-ai-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3,<3",
    "httpx>=0.27,<1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
